+++ docs/changelog.rst
@@ -44,7 +44,7 @@ a. The :py:func:`pingouin.homoscedasticity` gave WRONG results for wide-format d
 a. Partial correlation functions (:py:func:`pingouin.pcorr` and :py:func:`pingouin.partial_corr`) now use :py:func:`numpy.linalg.pinv` with `hermitian=True`, which improves numerical stability. See `issue 198 <https://github.com/raphaelvallat/pingouin/issues/198>`_.
 b. Disabled dtype check in :py:func:`pingouin.partial_corr` — previously, Pingouin could raise an error if the column names were integers. See `issue 201 <https://github.com/raphaelvallat/pingouin/issues/201>`_.
 c. :py:func:`pingouin.pairwise_corr` now works when the column names of the dataframe are integer, and better support numpy.arrays in the ``columns`` argument.
-d. Added support for wide-format dataframe in :py:func:`pingouin.friedman`
+d. Added support for wide-format dataframe in :py:func:`pingouin.friedman` and :py:func:`pingouin.cochran`
 
 v0.4.0 (August 2021)
 --------------------