+++ docs/changelog.rst
@@ -8,6 +8,16 @@ What's new
 
 *************
 
+v0.6.0.dev
+----------
+
+**Enhancements**
+
+a. Faster implementation of :py:func:`pingouin.gzscore`, adding all options available in zscore: axis, ddof and nan_policy. Warning: this functions is deprecated and will be removed in pingouin 0.7.0 (use scipy.stats.gzscore instead). See `pull request 210 <https://github.com/raphaelvallat/pingouin/pull/210>`_.
+
+
+*************
+
 v0.5.0 (October 2021)
 ---------------------